requirements.txt
python-telegram-bot[job-queue]==20.7
qrcode[pil]==7.4.2
Pillow==10.2.0
python-dateutil==2.8.2
pymongo[srv]==4.6.1
dnspython==2.6.1
aiohttp>=3.9.0
python-dotenv==1.0.1
boto3==1.34.84
